Medical Record Fix
-textlimit to 30
-functional version
