+++ app/src/main/res/layout/add_medicalrecord.xml
@@ -46,6 +46,7 @@
         android:layout_below="@+id/radioGroup"
         android:ems="10"
         android:hint="Enter the name here!"
+        android:maxLength="30"
         android:inputType="text" />
 
     <Button
@@ -65,7 +66,7 @@
         android:layout_alignParentStart="true"
         android:layout_below="@+id/button_addMR_dialog"
         android:background="@android:color/holo_blue_light"
-        android:text="Cancel" />
+        android:text="Done" />
 
     <TextView
         android:id="@+id/tv_title"
